Rename build_grid.py to a more descriptive name.
I'm keeping the old build_grid.py file until the autobumper config is updated to use this new filename as well.
