+++ config/jobs/kubernetes/kops/Makefile
@@ -14,4 +14,4 @@
 
 .PHONY: generate
 generate:
-	python3 build_grid.py
+	python3 build_jobs.py
